+++ src/mariadb/query.c
@@ -8,6 +8,15 @@ enum nss_status maria_init_db_conn(Maria_config *settings, MYSQL **conn, int *er
     return NSS_STATUS_TRYAGAIN;
   }
 
+  if(mysql_optionsv(*conn, MYSQL_READ_DEFAULT_FILE, (void *)settings->mariadb_client_config)) {
+    maria_log("Cannot load mariadb client options");
+    log_mysql_error(*conn);
+    *errnop = ENOENT;
+    return NSS_STATUS_UNAVAIL;
+  } else {
+    debug_print("custom options file loaded");
+  };
+
   if(mysql_real_connect(
     *conn,
     settings->dbhost,
@@ -19,13 +28,14 @@ enum nss_status maria_init_db_conn(Maria_config *settings, MYSQL **conn, int *er
     0
   ) == NULL) {
     maria_log("cannot connect to the database");
-    maria_log("settings;dbhost:%s;dbname:%s;db(root)user:%s;db(root)pass:%s;dbport:%lld;unix_socket:%s",
+    maria_log("settings;dbhost:%s;dbname:%s;db(root)user:%s;db(root)pass:%s;dbport:%lld;unix_socket:%s;mariadb_client_config:%s",
       settings->dbhost,
       settings->dbname,
       use_root_user ? settings->dbrootuser : settings->dbuser,
       use_root_user ? settings->dbrootpass : settings->dbpass,
       settings->dbport,
-      settings->unix_socket
+      settings->unix_socket,
+      settings->mariadb_client_config
     );
     log_mysql_error(*conn);
     *errnop = EAGAIN;